server/requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
docopt==0.6.2
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
geopandas==1.1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
networkx==3.5
numpy==2.3.4
osmnx==2.0.6
overpy==0.7
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pipreqs==0.4.13
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
scipy==1.16.2
shapely==2.1.2
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yarg==0.1.10
